02_MLOps/wine-quality-mlflow-e2e/requirements.txt
pandas
mlflow==2.2.2
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
PyYAML
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
-e .
